feat(shiftCell): only admin can manipulate shiftCell - regular user can't do anything but view, not even drag

+++ frontend/src/Dashboard/Slot/Shift/ShiftCell.tsx
@@ -1,12 +1,13 @@
 import { Flex, Text } from "@mantine/core";
 
 import {DateTime} from 'luxon'
-import {  useState } from "react";
+import {  useContext, useState } from "react";
 import TimeRangePicker from "../TimeRangePicker/TimeRangePicker";
 
 import axios from "axios";
 import type { Shift } from "../../Interfaces/Shift";
 import { useDraggable } from "@dnd-kit/core";
+import { AuthenticatedUser } from "../../../AuthenticatedUserContext";
 
 
 
@@ -19,6 +20,8 @@ const convertTo12hr = (time:string | null)=>{
 
 function ShiftCell({shift}:{shift:Shift}){
 
+    const isUserAdmin = useContext(AuthenticatedUser)?.role === 'ADMIN'
+    console.log(isUserAdmin)
     const start_time = convertTo12hr(shift!.start_time)
     const end_time = convertTo12hr(shift!.end_time)
 
@@ -51,13 +54,14 @@ function ShiftCell({shift}:{shift:Shift}){
     return (
     
         <Flex 
-            ref={setNodeRef} style={style} {...listeners} {...attributes}
-
-
-
+            ref={isUserAdmin ? setNodeRef : undefined}   
+            {...(isUserAdmin ? listeners : {})}         
+            {...(isUserAdmin ? attributes : {})}   
+            style={style}
             className="slot" 
             tabIndex={0} flex={1}  
             onKeyDown={(e)=>{
+                if(!isUserAdmin) return
                 if(e.key ==='Backspace' || e.key ==='Delete'){
                     console.log('Delete')
                     deleteShift()
@@ -66,11 +70,10 @@ function ShiftCell({shift}:{shift:Shift}){
                     console.log('space')
                     setActivate(prev=>!prev)
                 }
-
-
             }}
             onClick={()=>{
-                 setActivate(true)
+                if(!isUserAdmin) return
+                setActivate(true)
 
 
             }}
